% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackbox.R
\name{black_box}
\alias{black_box}
\title{Black Box}
\usage{
black_box(series, l, r = NULL, s = NULL)
}
\arguments{
\item{series}{Vector or Matrix of the time series to black-box.}

\item{l}{Numeric giving the number of sources in the collection.}

\item{r}{Vector giving the history lengths.}

\item{s}{Vector giving the future lengths.}
}
\value{
Vector giving the black-boxed time series.
}
\description{
Black-box time series from a collection of \code{l} sources where each
series has the same number of time steps, same number of initial conditions,
and possibly different bases into a single time series. The base of the
resulting time series is given by the product of the bases of each time
series in the collection. Black-boxing can be performed in time by
providing history lengths \code{r} and future lengths through \code{s}.
}
