% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{copy}
\alias{copy}
\title{Copy}
\usage{
copy(d)
}
\arguments{
\item{d}{Dist object representing the distribution to be copied.}
}
\value{
Dist giving the copy of the distribution.
}
\description{
Generic function to perform a deep copy of the distribution.
}
\examples{
d <- Dist(c(1, 2, 3, 4))
c <- copy(d)
c

d <- Dist(7)
c <- copy(d)
c
}
