% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{counts}
\alias{counts}
\title{Counts}
\usage{
counts(d)
}
\arguments{
\item{d}{Dist object representing the distribution.}
}
\value{
Numeric giving the number of observations.
}
\description{
Generic function to return the number of observations made thus far.
}
\examples{
d <- Dist(5)
counts(d) # 0

d <- Dist(c(1, 0, 3, 2))
counts(d) # 6
}
