% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dump}
\alias{dump}
\title{Dump}
\usage{
dump(d)
}
\arguments{
\item{d}{Dist object representing the distribution.}
}
\value{
Vector giving the empirical probabilities of all events.
}
\description{
Generic function to compute the empirical probability of each observable
event and return the result as an array.
}
\examples{
d <- Dist(c(0, 1, 2, 3, 4, 5, 6))
dump(d)
}
