% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{get_item}
\alias{get_item}
\title{Get Item}
\usage{
get_item(d, event)
}
\arguments{
\item{d}{Dist object representing the distribution.}

\item{event}{Numeric representing the observed event.}
}
\value{
Numeric giving the number of observation of \code{event}.
}
\description{
Generic function giving the number of observations made of \code{event}.
}
\examples{
d <- Dist(c(5, 0, 3, 7, 18, 0))
get_item(d, 1) # 5
get_item(d, 4) # 7
}
