% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{is_not_corrupted}
\alias{is_not_corrupted}
\title{Validity of the Dist object}
\usage{
is_not_corrupted(d)
}
\arguments{
\item{d}{Dist object representing the distribution.}
}
\value{
Logical signifying that the object is not corrupted.
}
\description{
Checks if the object \code{d} is not corrupted.
}
\keyword{internal}
