% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{partitioning}
\alias{partitioning}
\title{Partitioning}
\usage{
partitioning(n)
}
\arguments{
\item{n}{Numeric specifying the number of items of the set to partition.}
}
\value{
Matrix giving all possible partitioning schemata.
}
\description{
Partition a number \code{n} of items into all possible different subsets.
The number of partitions is given by the Bell number `B_n'.
}
\examples{
# All possible partitions of a set with 2 items
partitioning(2)

# All possible partitions of a set with 4 items
P <- partitioning(4)
t(P)
}
