% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{probability}
\alias{probability}
\title{Probability}
\usage{
probability(d, event)
}
\arguments{
\item{d}{Dist object representing the distribution.}

\item{event}{Numeric representing the observed event.}
}
\value{
Numerical giving the empirical probability of \code{event}.
}
\description{
Generic function to compute the empirical probability of an \code{event}.
}
\examples{
d <- Dist(c(0, 1, 2, 3, 4, 5))
for (i in 1:length(d))
  cat("event", i, "probability", probability(d, i), "\\n")
}
