% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{set_item}
\alias{set_item}
\title{Set Item}
\usage{
set_item(d, event, value)
}
\arguments{
\item{d}{Dist object representing the distribution.}

\item{event}{Numeric representing the observed event.}

\item{value}{Numeric representing the number of observations.}
}
\value{
Dist giving the modified distribution.
}
\description{
Generic function setting the number of observations made of \code{event}.
}
\examples{
d <- Dist(5)
d <- set_item(d, 1, 5)
d <- set_item(d, 3, 2)
d <- set_item(d, 5, 8)
d
}
