% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{tick}
\alias{tick}
\title{Tick}
\usage{
tick(d, event)
}
\arguments{
\item{d}{Dist object representing the distribution.}

\item{event}{Numeric representing the observed event.}
}
\value{
Dist giving the updated distribution.
}
\description{
Generic function to make a single observation of \code{event}, and return
the a modified distribution with the updated number of observations of said
\code{event}.
}
\examples{
d <- Dist(5)
d <- tick(d, 1)
d <- tick(d, 1)
d <- tick(d, 3)
d # [2, 0, 1, 0]

}
