% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{valid}
\alias{valid}
\title{Valid}
\usage{
valid(d)
}
\arguments{
\item{d}{Dist object representing the distribution.}
}
\value{
Logical representing the validity of the distribution.
}
\description{
Generic function to determine if the distribution is a valid probability
distribution, i.e. if the support is not empty and at least one observation
has been made.
}
\examples{
d <- Dist(5)
valid(d) # FALSE, support is empty

d <- Dist(c(0, 1, 0))
valid(d) # TRUE, at least one observation is made
}
